os/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database & Storage
supabase==2.9.1

# Redis & Task Queue
redis==5.0.1
upstash-redis==0.15.0

# AI & LLM
openai==1.54.0
anthropic==0.39.0

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2
httpx>=0.24.0